correct opening and closing of mspace; ensure correct calculation of global node identifier from population offsets; validated round-trip import and export of connectivity

+++ reader/ngh5types.hh
@@ -6,7 +6,7 @@
 #include "hdf5.h"
 
 // Block offset type
-typedef uint32_t DST_BLK_PTR_T;
+typedef uint64_t DST_BLK_PTR_T;
 
 // DBS offset type
 typedef uint64_t DST_PTR_T;
